+++ client/src/components/Content.jsx
@@ -1,0 +1,49 @@
+import React from 'react';
+
+import { makeStyles } from '@material-ui/core/styles';
+import { FormControl, InputLabel, MenuItem, Select } from '@material-ui/core';
+
+const useStyles = makeStyles((theme) => ({
+  formControl: {
+    margin: theme.spacing(1),
+    minWidth: 200,
+  },
+  selectEmpty: {
+    marginTop: theme.spacing(2),
+  },
+}));
+
+const Content = () => {
+  const classes = useStyles();
+
+  return (
+    <div>
+      <FormControl className={classes.formControl}>
+        <InputLabel id="select_skin">Select Skin Type</InputLabel>
+        <Select labelId="select_skin">
+          <MenuItem value={10}>Ten</MenuItem>
+          <MenuItem value={20}>Twenty</MenuItem>
+          <MenuItem value={30}>Thirty</MenuItem>
+        </Select>
+      </FormControl>
+      <FormControl className={classes.formControl}>
+        <InputLabel id="select_location">Select Location</InputLabel>
+        <Select labelId="select_location">
+          <MenuItem value={10}>Ten</MenuItem>
+          <MenuItem value={20}>Twenty</MenuItem>
+          <MenuItem value={30}>Thirty</MenuItem>
+        </Select>
+      </FormControl>
+      <FormControl className={classes.formControl}>
+        <InputLabel id="select_time">Select Time of Day</InputLabel>
+        <Select labelId="select_time">
+          <MenuItem value={10}>Ten</MenuItem>
+          <MenuItem value={20}>Twenty</MenuItem>
+          <MenuItem value={30}>Thirty</MenuItem>
+        </Select>
+      </FormControl>
+    </div>
+  );
+};
+
+export default Content;
