+++ client/src/components/NavBar.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './navbar.css';
+
+const NavBar = () => (
+  <div className="navbar">
+    <h2>Will I Tan?</h2>
+  </div>
+);
+
+export default NavBar;
