+++ client/src/components/navbar.css
@@ -1,0 +1,5 @@
+.navbar {
+  display: flex;
+  justify-content: center;
+  background-color: grey;
+}
